refactor: 상품 삭제 존재 여부 검사 분리

+++ server/src/services/products.service.ts
@@ -1,5 +1,5 @@
 import { CreateProductDto } from "../../interfaces/product.interface.js";
-import { save, findAll, deleteById } from "../repositories/products.repository.js";
+import { save, findAll, deleteById, isAlreadyExist } from "../repositories/products.repository.js";
 
 export async function addProduct(product: CreateProductDto) {
   await save(product);
@@ -10,5 +10,9 @@ export async function getProducts() {
 }
 
 export async function deleteProduct(id: number): Promise<boolean> {
-  return await deleteById(id);
+  if (!isAlreadyExist(id)) {
+    return false;
+  }
+  await deleteById(id);
+  return true;
 }
